Move Coroutine into a separate file
