+++ tests/unit/util/CancellationPlayground.cpp
@@ -31,101 +31,6 @@ print(char const* s, Args&&... args)
 }
 
 struct CancellationPlaygroundTest : SyncAsioContextTest {
-    class Coroutine {
-    public:
-        using cancellable_yield_context_type =
-            boost::asio::cancellation_slot_binder<boost::asio::yield_context, boost::asio::cancellation_slot>;
-
-    private:
-        boost::asio::yield_context yield_;
-        boost::system::error_code error_;
-        boost::asio::cancellation_signal cancellationSignal_;
-        cancellable_yield_context_type cyield_;
-        size_t generation_;
-
-        using GlobalSignal = boost::signals2::signal<void(size_t, boost::asio::cancellation_type_t)>;
-        std::shared_ptr<GlobalSignal> signal_;
-        boost::signals2::connection connection_;
-
-        explicit Coroutine(
-            boost::asio::yield_context&& yield,
-            std::shared_ptr<GlobalSignal> signal = std::make_shared<GlobalSignal>(),
-            size_t generation = 0
-        )
-            : yield_(std::move(yield))
-            , cyield_(boost::asio::bind_cancellation_slot(cancellationSignal_.slot(), yield_))
-            , generation_{generation}
-            , signal_{std::move(signal)}
-            , connection_{signal_->connect(
-                  [this](size_t generationToCancel, boost::asio::cancellation_type_t cancellationType) {
-                      if (generation_ >= generationToCancel) {
-                          cancellationSignal_.emit(cancellationType);
-                      }
-                  }
-              )}
-
-        {
-        }
-
-    public:
-        ~Coroutine()
-        {
-            // Could have used scoped_connection, but it is better to explicitly show the disconnect
-            connection_.disconnect();
-        }
-
-        template <typename ExecutionContext, std::invocable<Coroutine&> Fn>
-        static void
-        spawnNew(ExecutionContext& ioContext, Fn&& fn)
-        {
-            boost::asio::spawn(ioContext, [fn = std::forward<Fn>(fn)](boost::asio::yield_context yield) {
-                Coroutine thisCoroutine{std::move(yield)};
-                fn(thisCoroutine);
-            });
-        }
-
-        template <std::invocable<Coroutine&> Fn>
-        void
-        spawnChild(Fn&& fn)
-        {
-            boost::asio::spawn([nextGeneration = generation_ + 1,
-                                signal = signal_,
-                                fn = std::forward<Fn>(fn)](boost::asio::yield_context yield) mutable {
-                Coroutine coroutine(std::move(yield), std::move(signal), nextGeneration);
-                fn(coroutine);
-            });
-        }
-
-        boost::system::error_code
-        error() const
-        {
-            return error_;
-        }
-
-        void
-        cancel(boost::asio::cancellation_type_t cancellationType = boost::asio::cancellation_type::terminal)
-        {
-            signal_->operator()(generation_, cancellationType);
-        }
-        void
-        cancelAll(boost::asio::cancellation_type_t cancellationType = boost::asio::cancellation_type::terminal)
-        {
-            signal_->operator()(0, cancellationType);
-        }
-
-        bool
-        isCancelled() const
-        {
-            return error_ == boost::asio::error::operation_aborted;
-        }
-
-        cancellable_yield_context_type
-        yieldContext() const
-        {
-            return cyield_;
-        }
-    };
-
     size_t
     doWork(Coroutine& coroutine)
     {
@@ -161,15 +66,9 @@ struct CancellationPlaygroundTest : SyncAsioContextTest {
 
         print("starting asyncOperation {}", id);
 
-        if constexpr (requires { yield.get(); }) {
-            auto ioContext = yield.get().get_executor();
-            boost::asio::steady_timer timer{ioContext, std::chrono::milliseconds{durationMs}};
-            timer.async_wait(yield);
-        } else {
-            auto ioContext = yield.get_executor();
-            boost::asio::steady_timer timer{ioContext, std::chrono::milliseconds{durationMs}};
-            timer.async_wait(yield);
-        }
+        auto ioContext = yield.get().get_executor();
+        boost::asio::steady_timer timer{ioContext, std::chrono::milliseconds{durationMs}};
+        timer.async_wait(yield);
 
         std::println("{:%S}: finished asyncOperation {}", std::chrono::system_clock::now(), id);
     }
